Move m_floatingObjects to RenderBlockFlow from RenderBlock
Floats are part of block flow, not every RenderBlock can contain
floats. This change makes it so that RenderBlock no longer can contain
floats, only RenderBlockFlow can.

This also moves a bunch of line layout stuff over to RenderBlockFlow,
but makes no effort to move all of line layout, as all of that is
destined to move in future refactorings.

Review URL: https://codereview.chromium.org/27518011

git-svn-id: svn://svn.chromium.org/blink/trunk@160292 bbb929c8-8fbe-4397-9dbb-9b2b20218538