Move float logical location/dimension methods to RenderBlockFlow
I recently moved the logical location and dimension methods from RenderBlock to
FloatingObject. This made it so that the methods needed to take in the
horizontal writing mode of the FloatingObject's container in order to do their
thing. When looking into how to fix this, I came to understand why RenderBlock
has methods to determine the logical location and dimensions of its children,
when the children themselves have those methods: the methods on RenderBlock are
for determining the logical value in the writing mode of that RenderBlock, and
the method on the child is for determining the logical value in the writing
mode of the child.

Since the floating object list has moved from RenderBlock to RenderBlockFlow,
this change moves the logical dimension methods from FloatingObject to
RenderBlockFlow.

Review URL: https://codereview.chromium.org/37443002

git-svn-id: svn://svn.chromium.org/blink/trunk@160378 bbb929c8-8fbe-4397-9dbb-9b2b20218538