Refactor and clean up testharnessreport.js.
This change updates the content of testharnessreport.js to make it more
readable and conformant to the Blink style.

This commit does not change the behavior of testharness.js except for one
small difference. Previous testharnessreport.js prints the text output
using Element.innerHTML(), which makes the actual output broken if it contains
text surrounded with "<" and ">". This change fixes this issue, and updates
the expected results for the tests affected.

BUG=none
R=dpranke@chromium.org

Review URL: https://codereview.chromium.org/38673006

git-svn-id: svn://svn.chromium.org/blink/trunk@160508 bbb929c8-8fbe-4397-9dbb-9b2b20218538