Encapsulate the RenderStyle related checks in SharedStyleFinder
This moves the checks related to the state of RenderStyle for style sharing
into a method on RenderStyle (isSharable) in anticipation of moving the
checks into recalcStyle and attach to avoid having to check candidates that
will never allow sharing.

This patch does not change any behavior.

Review URL: https://codereview.chromium.org/49153004

git-svn-id: svn://svn.chromium.org/blink/trunk@160740 bbb929c8-8fbe-4397-9dbb-9b2b20218538