Don't allow style sharing of customized styles
There were many users of customStyleForRenderer() that were creating a
custom style and then not marking it as unique. This normally wasn't
an issue since the elements don't tend to be siblings of each other or
at the same level of the tree, but could manifest in future form control
types or the PasswordGeneratorButtonElement if we moved the calls to
addToStyleSharingList (like future patches might).

This patch adds all the missing setUnique() calls, but we should probably
consider some more general solution in the future or get rid of the
customStyleForRenderer() hook which is being used in bad ways all over
the UA shadow elements.

Review URL: https://codereview.chromium.org/47303006

git-svn-id: svn://svn.chromium.org/blink/trunk@160764 bbb929c8-8fbe-4397-9dbb-9b2b20218538