Avoid leaking MessagePortChannels when ports are cloned
MessagePortChannel used to hold a RefPtr to its entangled pair for the
sole purpose of answering MessagePortChannel::isConnectedTo(MessagePort*).
But answering that question only requires a pointer to the MessagePort, not
the channel.

This patch switches MessagePortChannel back to being non-refcounted, as it
was back when PlatformMessagePortChannel still existed. Neutered MessagePorts
thus delete their MessagePortChannel rather than simply deref()-ing them.

In the long run, getting rid of MessagePortChannel altogether and having
MessagePort directly interact with WebMessagePortChannel is probably the
right thing, but an incremental approach is likely the right route.

R=abarth
BUG=311229

Review URL: https://codereview.chromium.org/48793004

git-svn-id: svn://svn.chromium.org/blink/trunk@160798 bbb929c8-8fbe-4397-9dbb-9b2b20218538