Don't null the RenderView during document destruction
Instead of nulling out the RenderView when we're going to
destroy the Document, we explicitly check the document's
lifecycle state for Stopping.

I tried this once but it was reverted because recalcStyle()
and updateLayout() were not updated to avoid rendering layout
or recalcStyle when the render tree is being torn down.

Review URL: https://codereview.chromium.org/54373002

git-svn-id: svn://svn.chromium.org/blink/trunk@161121 bbb929c8-8fbe-4397-9dbb-9b2b20218538