[2D Canvas] Fix crash in createPattern method when using invalid canvas as source.
Before this fix, createPattern would crash when the source image is a canvas
element with a backing store that failed to allocate. Correct behavior
is to default to a transparent black pattern. The code to handle the
default behavior was already there, but the constructor of Pattern was
crashing when passed a null image.  This patch fixes that crash.

BUG=281504

Review URL: https://codereview.chromium.org/57643005

git-svn-id: svn://svn.chromium.org/blink/trunk@161286 bbb929c8-8fbe-4397-9dbb-9b2b20218538