IDL compiler: Update callback interfaces and start testing
Callback interfaces were not being tested, b/c Perl and Python output
were not exactly equal (due to code cleanup in Python that wasn't ported
back to Perl).
Thus the code bitrotted as we've changed the Perl and Python.
This makes the code work again, fixes all discrepencies, and turns the
test on so this doesn't happen in future.

Perl changes are whitespace and comments,
Python and template are mostly cleanup,
but some substantive change ({creation_context}).

BUG=239771

Review URL: https://codereview.chromium.org/54903006

git-svn-id: svn://svn.chromium.org/blink/trunk@161307 bbb929c8-8fbe-4397-9dbb-9b2b20218538