CSS Animations/Transitions: Step immediately when blending lengths that can't be smoothly interpolated
The existing transition engine incorrectly starts transitions between values
that can't be smoothly interpolated. After r159314 this meant that
interpolating a length eg. from '0px' to 'auto' would produce '0px' at the
beginning of the animation. This caused regressions when authors attempted to
read back the computed style immediately.

Modifing the existing transition engine to not start transitions in this case
is non-trivial, so this change modifies blending such that when interpolating
with an undefined, intrinsic or auto value we step immediately. After the
web-animations engine is enabled such transitions will not be started and we
can modify this logic to step at 50%.

BUG=313649

Review URL: https://codereview.chromium.org/63323003

git-svn-id: svn://svn.chromium.org/blink/trunk@161503 bbb929c8-8fbe-4397-9dbb-9b2b20218538